Removed backlit computations.
