+++ source/Core/Castor3D/Src/Shader/Shaders/GlslMetallicBrdfLighting.cpp
@@ -1,4 +1,4 @@
-﻿#include "GlslMetallicBrdfLighting.hpp"
+#include "GlslMetallicBrdfLighting.hpp"
 
 #include "GlslMaterial.hpp"
 #include "GlslShadow.hpp"
@@ -141,52 +141,12 @@ namespace castor3d
 			m_writer << Endi();
 		}
 
-		Vec3 MetallicBrdfLightingModel::computeBackLit( DirectionalLight const & light
-			, Vec3 const & worldEye
-			, Vec3 const & albedo
-			, Float const & metallic
-			, FragmentInput const & fragmentIn )const
-		{
-			return m_computeDirectionalBackLit( DirectionalLight{ light }
-				, worldEye
-				, albedo
-				, metallic
-				, FragmentInput{ fragmentIn } );
-		}
-
-		Vec3 MetallicBrdfLightingModel::computeBackLit( PointLight const & light
-			, Vec3 const & worldEye
-			, Vec3 const & albedo
-			, Float const & metallic
-			, FragmentInput const & fragmentIn )const
-		{
-			return m_computePointBackLit( PointLight{ light }
-				, worldEye
-				, albedo
-				, metallic
-				, FragmentInput{ fragmentIn } );
-		}
-
-		Vec3 MetallicBrdfLightingModel::computeBackLit( SpotLight const & light
-			, Vec3 const & worldEye
-			, Vec3 const & albedo
-			, Float const & metallic
-			, FragmentInput const & fragmentIn )const
-		{
-			return m_computeSpotBackLit( SpotLight{ light }
-				, worldEye
-				, albedo
-				, metallic
-				, FragmentInput{ fragmentIn } );
-		}
-
 		void MetallicBrdfLightingModel::doDeclareModel()
 		{
 			doDeclareDistribution();
 			doDeclareGeometry();
 			doDeclareFresnelShlick();
 			doDeclareComputeLight();
-			doDeclareComputeLightBackLit();
 		}
 
 		void MetallicBrdfLightingModel::doDeclareComputeDirectionalLight()
@@ -524,126 +484,6 @@ namespace castor3d
 				, output );
 		}
 
-		void MetallicBrdfLightingModel::doDeclareComputeDirectionalLightBackLit()
-		{
-			OutputComponents output{ m_writer };
-			m_computeDirectionalBackLit = m_writer.implementFunction< Vec3 >( cuT( "computeBackLit" )
-				, [this]( DirectionalLight const & light
-					, Vec3 const & worldEye
-					, Vec3 const & albedo
-					, Float const & metallic
-					, FragmentInput const & fragmentIn )
-				{
-					PbrMRMaterials materials{ m_writer };
-					auto lightDirection = m_writer.declLocale( cuT( "lightDirection" )
-						, normalize( -light.m_direction().xyz() ) );
-
-					m_writer.returnStmt( doComputeLightBackLit( light.m_lightBase()
-						, worldEye
-						, lightDirection
-						, albedo
-						, metallic
-						, fragmentIn ) );
-				}
-				, DirectionalLight( &m_writer, cuT( "light" ) )
-				, InVec3( &m_writer, cuT( "worldEye" ) )
-				, InVec3( &m_writer, cuT( "albedo" ) )
-				, InFloat( &m_writer, cuT( "metallic" ) )
-				, FragmentInput{ m_writer } );
-		}
-
-		void MetallicBrdfLightingModel::doDeclareComputePointLightBackLit()
-		{
-			OutputComponents output{ m_writer };
-			m_computePointBackLit = m_writer.implementFunction< Vec3 >( cuT( "computeBackLit" )
-				, [this]( PointLight const & light
-					, Vec3 const & worldEye
-					, Vec3 const & albedo
-					, Float const & metallic
-					, FragmentInput const & fragmentIn )
-				{
-					PbrMRMaterials materials{ m_writer };
-					auto lightToVertex = m_writer.declLocale( cuT( "lightToVertex" )
-						, light.m_position().xyz() - fragmentIn.m_vertex );
-					auto distance = m_writer.declLocale( cuT( "distance" )
-						, length( lightToVertex ) );
-					auto lightDirection = m_writer.declLocale( cuT( "lightDirection" )
-						, normalize( lightToVertex ) );
-
-					auto backLit = m_writer.declLocale( cuT( "backLit" )
-						, doComputeLightBackLit( light.m_lightBase()
-							, worldEye
-							, lightDirection
-							, albedo
-							, metallic
-							, fragmentIn ) );
-					auto attenuation = m_writer.declLocale( cuT( "attenuation" )
-						, glsl::fma( light.m_attenuation().z()
-							, distance * distance
-							, glsl::fma( light.m_attenuation().y()
-								, distance
-								, light.m_attenuation().x() ) ) );
-					m_writer.returnStmt( backLit / attenuation );
-				}
-				, PointLight( &m_writer, cuT( "light" ) )
-				, InVec3( &m_writer, cuT( "worldEye" ) )
-				, InVec3( &m_writer, cuT( "albedo" ) )
-				, InFloat( &m_writer, cuT( "metallic" ) )
-				, FragmentInput{ m_writer } );
-		}
-
-		void MetallicBrdfLightingModel::doDeclareComputeSpotLightBackLit()
-		{
-			OutputComponents output{ m_writer };
-			m_computeSpotBackLit = m_writer.implementFunction< Vec3 >( cuT( "computeBackLit" )
-				, [this]( SpotLight const & light
-					, Vec3 const & worldEye
-					, Vec3 const & albedo
-					, Float const & metallic
-					, FragmentInput const & fragmentIn )
-				{
-					PbrMRMaterials materials{ m_writer };
-					auto lightToVertex = m_writer.declLocale( cuT( "lightToVertex" )
-						, light.m_position().xyz() - fragmentIn.m_vertex );
-					auto distance = m_writer.declLocale( cuT( "distance" )
-						, length( lightToVertex ) );
-					auto lightDirection = m_writer.declLocale( cuT( "lightDirection" )
-						, normalize( lightToVertex ) );
-					auto spotFactor = m_writer.declLocale( cuT( "spotFactor" )
-						, dot( lightDirection, -light.m_direction() ) );
-					auto backLit = m_writer.declLocale( cuT( "backLit" )
-						, vec3( 0.0_f ) );
-
-					IF( m_writer, spotFactor > light.m_cutOff() )
-					{
-						backLit = doComputeLightBackLit( light.m_lightBase()
-							, worldEye
-							, lightDirection
-							, albedo
-							, metallic
-							, fragmentIn );
-						auto attenuation = m_writer.declLocale( cuT( "attenuation" )
-							, glsl::fma( light.m_attenuation().z()
-								, distance * distance
-								, glsl::fma( light.m_attenuation().y()
-									, distance
-									, light.m_attenuation().x() ) ) );
-						spotFactor = glsl::fma( m_writer.paren( 1.0_f - spotFactor )
-							, 1.0_f / m_writer.paren( 1.0_f - light.m_cutOff() )
-							, 1.0_f );
-						backLit = spotFactor * backLit / attenuation;
-					}
-					FI;
-
-					m_writer.returnStmt( backLit );
-				}
-				, SpotLight( &m_writer, cuT( "light" ) )
-				, InVec3( &m_writer, cuT( "worldEye" ) )
-				, InVec3( &m_writer, cuT( "albedo" ) )
-				, InFloat( &m_writer, cuT( "metallic" ) )
-				, FragmentInput{ m_writer } );
-		}
-
 		void MetallicBrdfLightingModel::doDeclareComputeLight()
 		{
 			OutputComponents output{ m_writer };
@@ -719,57 +559,6 @@ namespace castor3d
 				, InFloat( &m_writer, cuT( "shadowFactor" ) )
 				, FragmentInput{ m_writer }
 				, output );
-		}
-	
-		void MetallicBrdfLightingModel::doDeclareComputeLightBackLit()
-		{
-			OutputComponents output{ m_writer };
-			m_computeLightBackLit = m_writer.implementFunction< Vec3 >( cuT( "doComputeLightBackLit" )
-				, [this]( Light const & light
-					, Vec3 const & worldEye
-					, Vec3 const & direction
-					, Vec3 const & albedo
-					, Float const & metallic
-					, FragmentInput const & fragmentIn )
-				{
-					// From https://learnopengl.com/#!PBR/Lighting
-					auto constexpr PI = 3.1415926535897932384626433832795028841968;
-					auto L = m_writer.declLocale( cuT( "L" )
-						, normalize( direction ) );
-					auto V = m_writer.declLocale( cuT( "V" )
-						, normalize( normalize( worldEye - fragmentIn.m_vertex ) ) );
-					auto H = m_writer.declLocale( cuT( "H" )
-						, normalize( L + V ) );
-					auto N = m_writer.declLocale( cuT( "N" )
-						, normalize( fragmentIn.m_normal ) );
-					auto radiance = m_writer.declLocale( cuT( "radiance" )
-						, light.m_colour() );
-
-					auto NdotL = m_writer.declLocale( cuT( "NdotL" )
-						, max( 0.0_f, dot( N, L ) ) );
-					auto HdotV = m_writer.declLocale( cuT( "HdotV" )
-						, max( 0.0_f, dot( H, V ) ) );
-
-					auto f0 = m_writer.declLocale( cuT( "f0" )
-						, mix( vec3( 0.04_f ), albedo, metallic ) );
-					auto specfresnel = m_writer.declLocale( cuT( "specfresnel" )
-						, m_schlickFresnel( HdotV, f0 ) );
-			
-					auto kS = m_writer.declLocale( cuT( "kS" )
-						, specfresnel );
-					auto kD = m_writer.declLocale( cuT( "kD" )
-						, vec3( 1.0_f ) - kS );
-
-					kD *= 1.0_f - metallic;
-
-					m_writer.returnStmt( radiance * NdotL * kD / PI );
-				}
-				, InLight( &m_writer, cuT( "light" ) )
-				, InVec3( &m_writer, cuT( "worldEye" ) )
-				, InVec3( &m_writer, cuT( "direction" ) )
-				, InVec3( &m_writer, cuT( "albedo" ) )
-				, InFloat( &m_writer, cuT( "metallic" ) )
-				, FragmentInput{ m_writer } );
 		}
 
 		void MetallicBrdfLightingModel::doDeclareDistribution()
@@ -881,21 +670,6 @@ namespace castor3d
 				, FragmentInput{ fragmentIn }
 				, output );
 			m_writer << Endi();
-		}
-	
-		Vec3 MetallicBrdfLightingModel::doComputeLightBackLit( Light const & light
-			, Vec3 const & worldEye
-			, Vec3 const & direction
-			, Vec3 const & albedo
-			, Float const & metallic
-			, FragmentInput const & fragmentIn )const
-		{
-			return m_computeLightBackLit( light
-				, worldEye
-				, direction
-				, albedo
-				, metallic
-				, FragmentInput{ fragmentIn } );
 		}
 
 		//***********************************************************************************************
