+++ source/Core/Castor3D/Src/Shader/Shaders/GlslMetallicBrdfLighting.hpp
@@ -47,21 +47,6 @@ namespace castor3d
 				, glsl::Int const & receivesShadows
 				, FragmentInput const & fragmentIn
 				, OutputComponents & output )const;
-			C3D_API glsl::Vec3 computeBackLit( DirectionalLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & albedo
-				, glsl::Float const & metallic
-				, FragmentInput const & fragmentIn )const;
-			C3D_API glsl::Vec3 computeBackLit( PointLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & albedo
-				, glsl::Float const & metallic
-				, FragmentInput const & fragmentIn )const;
-			C3D_API glsl::Vec3 computeBackLit( SpotLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & albedo
-				, glsl::Float const & metallic
-				, FragmentInput const & fragmentIn )const;
 
 		protected:
 			void doDeclareModel();
@@ -70,9 +55,6 @@ namespace castor3d
 			void doDeclareComputeSpotLight()override;
 			void doDeclareComputeOnePointLight()override;
 			void doDeclareComputeOneSpotLight()override;
-			void doDeclareComputeDirectionalLightBackLit()override;
-			void doDeclareComputePointLightBackLit()override;
-			void doDeclareComputeSpotLightBackLit()override;
 
 			void doComputeLight( Light const & light
 				, glsl::Vec3 const & worldEye
@@ -84,18 +66,10 @@ namespace castor3d
 				, FragmentInput const & fragmentIn
 				, OutputComponents & output )const;
 
-			glsl::Vec3 doComputeLightBackLit( Light const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & direction
-				, glsl::Vec3 const & albedo
-				, glsl::Float const & metallic
-				, FragmentInput const & fragmentIn )const;
-
 			void doDeclareDistribution();
 			void doDeclareGeometry();
 			void doDeclareFresnelShlick();
 			void doDeclareComputeLight();
-			void doDeclareComputeLightBackLit();
 
 		public:
 			C3D_API static const castor::String Name;
@@ -178,31 +152,6 @@ namespace castor3d
 				, glsl::InFloat
 				, FragmentInput
 				, OutputComponents & > m_computeLight;
-			glsl::Function< glsl::Vec3
-				, DirectionalLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InFloat
-				, FragmentInput > m_computeDirectionalBackLit;
-			glsl::Function< glsl::Vec3
-				, PointLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InFloat
-				, FragmentInput > m_computePointBackLit;
-			glsl::Function< glsl::Vec3
-				, SpotLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InFloat
-				, FragmentInput > m_computeSpotBackLit;
-			glsl::Function< glsl::Vec3
-				, InLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InFloat
-				, FragmentInput > m_computeLightBackLit;
 		};
 	}
 }
