+++ source/Core/Castor3D/Src/Shader/Shaders/GlslSpecularBrdfLighting.hpp
@@ -47,18 +47,6 @@ namespace castor3d
 				, glsl::Int const & receivesShadows
 				, FragmentInput const & fragmentIn
 				, OutputComponents & output )const;
-			C3D_API glsl::Vec3 computeBackLit( DirectionalLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & specular
-				, FragmentInput const & fragmentIn )const;
-			C3D_API glsl::Vec3 computeBackLit( PointLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & specular
-				, FragmentInput const & fragmentIn )const;
-			C3D_API glsl::Vec3 computeBackLit( SpotLight const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & specular
-				, FragmentInput const & fragmentIn )const;
 
 		protected:
 			void doDeclareModel();
@@ -67,9 +55,6 @@ namespace castor3d
 			void doDeclareComputeSpotLight()override;
 			void doDeclareComputeOnePointLight()override;
 			void doDeclareComputeOneSpotLight()override;
-			void doDeclareComputeDirectionalLightBackLit()override;
-			void doDeclareComputePointLightBackLit()override;
-			void doDeclareComputeSpotLightBackLit()override;
 
 			void doComputeLight( Light const & light
 				, glsl::Vec3 const & worldEye
@@ -81,17 +66,10 @@ namespace castor3d
 				, FragmentInput const & fragmentIn
 				, OutputComponents & output );
 
-			glsl::Vec3 doComputeLightBackLit( Light const & light
-				, glsl::Vec3 const & worldEye
-				, glsl::Vec3 const & direction
-				, glsl::Vec3 const & specular
-				, FragmentInput const & fragmentIn );
-
 			void doDeclareDistribution();
 			void doDeclareGeometry();
 			void doDeclareFresnelShlick();
 			void doDeclareComputeLight();
-			void doDeclareComputeLightBackLit();
 
 		public:
 			C3D_API static const castor::String Name;
@@ -174,27 +152,6 @@ namespace castor3d
 				, glsl::InFloat
 				, FragmentInput
 				, OutputComponents & > m_computeLight;
-			glsl::Function< glsl::Vec3
-				, DirectionalLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, FragmentInput > m_computeDirectionalBackLit;
-			glsl::Function< glsl::Vec3
-				, PointLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, FragmentInput > m_computePointBackLit;
-			glsl::Function< glsl::Vec3
-				, SpotLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, FragmentInput > m_computeSpotBackLit;
-			glsl::Function< glsl::Vec3
-				, InLight
-				, glsl::InVec3
-				, glsl::InVec3
-				, glsl::InVec3
-				, FragmentInput > m_computeLightBackLit;
 		};
 	}
 }
